Report errors detected in SBML when parsing
Prior to this change, the checkSBMLDocument function in
SBMLLint/common/util.py would check if the error count during parsing
of the document was non-zero and raise a ValueError if so.  This
change still raises a ValueError, but it adds the contents of
the python-libsml error log to the ValueError to the message
parameter to enable users to identify and correct parsing errors.
In addition, this change modifies the lint function in
SBMLLint/tools/sbmllint.py to pass the name of the SBML file that
is currently being parsed to the call to checkSBMLDocument so that
the ValueError can report the model filename in the expression
parameter, rather than an empty string.